Prevent malformed response from onHeaders early return
Before this commit, a "Transfer-Encoding: chunked" request to a Shiny app would
be responded to with the following, in order:

1) Error 413 headers
2) Error 404 headers
3) Body for (2)
4) Body for (1)

This resulted in an unparseable mess for the client, which in the case of Shiny
Server, would cause the proxy to dispose itself and then a SIGPIPE to occur in
the R process.

With this fix, only (1) and (4) are sent, so everything is well-formed.