Fix split header bug
There was an error in the logic of the websocket frame parser, where
if the header was split over two separate TCP packets, the parser
would miscalculate the start of the body. Since client websocket
packets are XOR-ed with a masking key, this would result in the
message looking like total garbage.

Repro case was with Safari 7.0 running against RStudio Server Pro
with SSL enabled, running a Shiny app from the console.