Allow responses to omit a body and the Content-Length header.
RFC 7231 (part of the HTTP 1.1 specification) states[0]:

> A 204 response is terminated by the first empty line after the header
> fields because it cannot contain a message body.

while RFC 7232 says[1]:

> A 304 response cannot contain a message-body; it is always terminated
> by the first empty line after the header fields.

In addition, RFC 7230 says the following about the Content-Length
header[2]:

> A server MAY send a Content-Length header field in a 304 (Not
> Modified) response to a conditional GET request (Section 4.1 of
> [RFC7232]); a server MUST NOT send Content-Length in such a response
> unless its field-value equals the decimal number of octets that would
> have been sent in the payload body of a 200 (OK) response to the same
> request.
>
> A server MUST NOT send a Content-Length header field in any response
> with a status code of 1xx (Informational) or 204 (No Content).

That is, we need to provide a way to suppress the inclusion of a body
and the generated Content-Length header so that downstream users can
emit correct HTTP 204 and 304 responses. This commit makes that possible
by allowing the Rook response to have a missing (or NULL) body, in which
case the usual body handling will be skipped.

It's still possible for users to add a Content-Length header directly,
which they might want to do, e.g. to fulfill the HTTP 304 behaviour
quoted above.

As per Winston Chang's request, we don't explicitly handle these status
codes -- users must implement this behaviour themselves.

Unit tests and updated documentation on the reponse object are included.

[0]: https://tools.ietf.org/html/rfc7231#section-6.3.5
[1]: https://tools.ietf.org/html/rfc7232#section-4.1
[2]: https://tools.ietf.org/html/rfc7230#section-3.3.2