Move examples to port 8080 and fix typo
Mac OS Monterey now reserves port `5000` for "AirPlay Receiver" which means the examples fail in a confusing and hard-to-debug way. 
Also fixes a small typo where `path = <PORT NUMBER>` was used. Removed the argument name to match the rest of the examples and so the port number given actually is used.